+++ app/src/main/java/com/jaychang/demo/srv/cell/BookCell.java
@@ -80,13 +80,13 @@ public class BookCell extends SimpleCell<Book, BookCell.ViewHolder>
     // do your cleaning jobs here when the item view is recycled.
   }
 
+  public void setShowHandle(boolean showHandle) {
+    this.showHandle = showHandle;
+  }
+
   @Override
   protected long getItemId() {
     return getItem().getId();
-  }
-
-  public void setShowHandle(boolean showHandle) {
-    this.showHandle = showHandle;
   }
 
   /**