+++ library/src/main/java/com/jaychang/srv/OnLoadMoreListener.java
@@ -1,5 +1,6 @@
 package com.jaychang.srv;
 
 public interface OnLoadMoreListener {
-  void onLoadMore(SimpleRecyclerView simpleRecyclerView);
+  boolean shouldLoadMore();
+  void onLoadMore();
 }