+++ library/src/main/java/com/jaychang/srv/decoration/SectionHeaderItemDecoration.java
@@ -113,6 +113,10 @@ public class SectionHeaderItemDecoration extends RecyclerView.ItemDecoration {
       return;
     }
 
+    if (layoutManager == null) {
+      layoutManager = (LinearLayoutManager) parent.getLayoutManager();
+    }
+
     int position = layoutManager.findFirstVisibleItemPosition();
 
     if (position == NO_POSITION) {
@@ -159,8 +163,8 @@ public class SectionHeaderItemDecoration extends RecyclerView.ItemDecoration {
 
   private View getAndMeasureSectionHeader(RecyclerView parent, int position) {
     View sectionHeader = provider.getSectionHeaderView(getItem(position), position);
-    int widthSpec = View.MeasureSpec.makeMeasureSpec(parent.getWidth(), View.MeasureSpec.UNSPECIFIED);
-    int heightSpec = View.MeasureSpec.makeMeasureSpec(parent.getHeight(), View.MeasureSpec.UNSPECIFIED);
+    int widthSpec = View.MeasureSpec.makeMeasureSpec(parent.getWidth(), View.MeasureSpec.EXACTLY);
+    int heightSpec = View.MeasureSpec.makeMeasureSpec(parent.getHeight(), View.MeasureSpec.AT_MOST);
     sectionHeader.measure(widthSpec, heightSpec);
     return sectionHeader;
   }